Add files via upload
This program simulates what happens when a ball is thrown at a particular angle and initial velocity. The Ball.java includes all private fields and all doubles. It initializes methods and sets horizontal and vertical distances with a set of horizontal and vertical speeds.
AkakpoAssign4.java declares double variables for each input value. It reads the angle, reads the initial velocity, height and time interval.